requirements.txt
llama-index==0.11.6
llama-index-llms-openai
llama-index-readers-file
llama-index-embeddings-openai
llama-index-llms-openai-like
openinference-instrumentation-llama-index>=2
arize-phoenix
python-dotenv
pandas
numpy
streamlit
streamlit-chat
yahoo-oauth
yahoo-fantasy-api 